Small refactor; Add new fields to device model;
