+++ constant/index.js
@@ -1,0 +1,22 @@
+const randWords = [
+  'name',
+  'sensor',
+  'event',
+  'status',
+  'panic',
+  'metric',
+  'type',
+  'model_name',
+  'model_ID',
+  'bodyColor',
+  'firmwareVersion',
+  'lastSyncTime',
+  'alias',
+  'deviceStatus',
+  'totalJobs',
+  'printTimeComplete',
+];
+
+module.exports = {
+  randWords,
+};
